requirements.txt
arrow==0.16.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.15.0
Werkzeug==1.0.1
